+++ src/pages/products.js
@@ -1,142 +1,43 @@
-import React, { useState } from "react";
-import { DivProducts } from "./styles";
+import { Button, ButtonGroup } from "@material-ui/core";
+import React from "react";
 import Lista from "../components/list";
-const productsArray = [
-  {
-    image: "/assets/racoes/golden.jpeg",
-    value: "R$ 122,00",
-    name: "Golden",
-  },
-
-  {
-    image: "/assets/racoes/sapeca.jpeg",
-    value: "R$ 100,00",
-    name: "Sapeca",
-  },
-
-  {
-    image: "/assets/racoes/pastel dog.jpeg",
-    value: "R$ 140,00",
-    name: "Pastel Dog",
-  },
-
-  {
-    image: "/assets/racoes/premier.jpeg",
-    value: "R$ 140,00",
-    name: "Premier",
-  },
-  {
-    image: "/assets/racoes/golden.jpeg",
-    value: "R$ 122,00",
-    name: "Golden",
-  },
-
-  {
-    image: "/assets/racoes/sapeca.jpeg",
-    value: "R$ 100,00",
-    name: "Sapeca",
-  },
-
-  {
-    image: "/assets/racoes/pastel dog.jpeg",
-    value: "R$ 140,00",
-    name: "Pastel Dog",
-  },
-
-  {
-    image: "/assets/racoes/premier.jpeg",
-    value: "R$ 140,00",
-    name: "Premier",
-  },
-  {
-    image: "/assets/racoes/golden.jpeg",
-    value: "R$ 122,00",
-    name: "Golden",
-  },
-
-  {
-    image: "/assets/racoes/sapeca.jpeg",
-    value: "R$ 100,00",
-    name: "Sapeca",
-  },
-
-  {
-    image: "/assets/racoes/pastel dog.jpeg",
-    value: "R$ 140,00",
-    name: "Pastel Dog",
-  },
-
-  {
-    image: "/assets/racoes/premier.jpeg",
-    value: "R$ 140,00",
-    name: "Premier",
-  },
-  {
-    image: "/assets/racoes/golden.jpeg",
-    value: "R$ 122,00",
-    name: "Golden",
-  },
-
-  {
-    image: "/assets/racoes/sapeca.jpeg",
-    value: "R$ 100,00",
-    name: "Sapeca",
-  },
-
-  {
-    image: "/assets/racoes/pastel dog.jpeg",
-    value: "R$ 140,00",
-    name: "Pastel Dog",
-  },
-
-  {
-    image: "/assets/racoes/premier.jpeg",
-    value: "R$ 140,00",
-    name: "Premier",
-  },
-];
-
-const itemsPerPage = 9;
+import { green, yellow } from "../ui/colors";
 
 export default function Products() {
-  const [datosFromApi, setDatosFromApi] = useState(productsArray);
-  console.log(productsArray);
-  const [items, setItems] = useState(
-    [...productsArray].splice(0, itemsPerPage)
-  );
-
-  const [currentPage, setCurrentPage] = useState(0);
-
-  const nextHandler = () => {
-    const totalItems = datosFromApi.length;
-
-    const nextPage = currentPage + 1;
-
-    const firstIndex = nextPage * itemsPerPage;
-
-    if (firstIndex === totalItems) return;
-
-    setItems([...datosFromApi].splice(firstIndex, itemsPerPage));
-    setCurrentPage(nextPage);
-  };
-
-  const prevHandler = () => {
-    const prevPage = currentPage - 1;
-
-    if (prevPage < 0) return;
-    const firstIndex = prevPage * itemsPerPage;
-    setItems([...datosFromApi].splice(firstIndex, itemsPerPage));
-    setCurrentPage(prevPage);
-  };
-
   return (
-    <DivProducts>
-      <Lista
-        currentPage={currentPage}
-        items={items}
-        nextHandler={nextHandler}
-        prevHandler={prevHandler}
-      />
-    </DivProducts>
+    <div
+      style={{
+        borderRadius: 10,
+        textAlign: "center",
+        backgroundColor: green,
+        padding: 10,
+      }}
+    >
+      {/* <Title>Produtos</Title> */}
+      <Lista />
+      <ButtonGroup style={{ marginBottom: 40 }} orientation="horizontal">
+        <Button
+          style={{
+            backgroundColor: "#fed227",
+            color: `${green}`,
+            fontWeight: 600,
+            marginRight: 20,
+          }}
+          variant="contained"
+        >
+          Anterior
+        </Button>
+        <Button
+          style={{
+            backgroundColor: "#fed227",
+            color: `${green}`,
+            fontWeight: 600,
+          }}
+          variant="contained"
+        >
+          Siguiente
+        </Button>
+      </ButtonGroup>
+    </div>
   );
 }
