+++ src/pages/styles.js
@@ -1,3 +1,4 @@
+import { Button } from "@material-ui/core";
 import styled from "styled-components";
 import { green, lightGray, black, white } from "../ui/colors";
 
@@ -13,13 +14,6 @@ export const Header = styled.div`
   flex-direction: row;
   background-color: ${green};
   align-items: center;
-`;
-
-export const DivProducts = styled.div`
-  border-radius: 10;
-  text-align: center;
-  background-color: ${green};
-  padding: 10;
 `;
 
 //Text
